Comments Improvement for Spam/Suspicious

+++ app/admin/dashboards.rb
@@ -43,25 +43,24 @@ ActiveAdmin::Dashboards.build do
   
   ### COMMENTS
 
-  section "Recent Comments", priority: 1 do
+  section "Approved Comments", priority: 1 do
     if controller.current_ability.can?(:read, Comment)
-      table_for Comment.recent(16) do
+      table_for Comment.recent(12) do
         column "User" do |t|
           if t.owner
-            link_to t.owner.title, admin_user_path(t.owner)
+            link_to t.owner.title, admin_user_path(t.owner), class: "suspicious_#{t.suspicious?} spam_#{t.marked_spam?}"
           else
             t.owner_id
           end
         end
         column "Message" do |t|
-          link_to t.body.truncate(60), admin_comment_path(t), title: t.body
+          link_to t.body.truncate(80), admin_comment_path(t), title: t.body, class: "suspicious_#{t.suspicious?} spam_#{t.marked_spam?}"
         end
         column "Item" do |t|
           if t.commentable && t.commentable_type == "Item"
-            link_to t.commentable.title.truncate(50), admin_item_path(t.commentable), title: t.commentable.title
+            link_to t.commentable.title.truncate(44), admin_item_path(t.commentable), title: t.commentable.title
           elsif t.commentable && t.commentable_type == "Comment"
-            link_to t.commentable.commentable.title.truncate(50), admin_item_path(t.commentable.commentable)
-            # link_to("Reply to: #{t.commentable.body.truncate(50)}", admin_comment_path(t.commentable))
+            link_to t.commentable.commentable.title.truncate(44), admin_item_path(t.commentable.commentable)
           end
         end
         column "Time" do |t|
@@ -74,24 +73,24 @@ ActiveAdmin::Dashboards.build do
     end
   end
 
-  section "Pending Comments", priority: 2 do
+  section "Suspicious Comments", priority: 1 do
     if controller.current_ability.can?(:read, Comment)
-      table_for Comment.as_spam(16) do
+      table_for Comment.suspicious(12) do
         column "User" do |t|
           if t.owner
-            link_to t.owner.title, admin_user_path(t.owner)
+            link_to t.owner.title, admin_user_path(t.owner), class: "suspicious_#{t.suspicious?} spam_#{t.marked_spam?}"
           else
             t.owner_id
           end
         end
         column "Message" do |t|
-          link_to t.body.truncate(50), admin_comment_path(t), title: t.body
+          link_to t.body.truncate(80), admin_comment_path(t), title: t.body, class: "suspicious_#{t.suspicious?} spam_#{t.marked_spam?}"
         end
         column "Item" do |t|
           if t.commentable && t.commentable_type == "Item"
-            link_to t.commentable.title.truncate(50), admin_item_path(t.commentable)
+            link_to t.commentable.title.truncate(44), admin_item_path(t.commentable), title: t.commentable.title
           elsif t.commentable && t.commentable_type == "Comment"
-            link_to t.commentable.commentable.title.truncate(50), admin_item_path(t.commentable.commentable)
+            link_to t.commentable.commentable.title.truncate(44), admin_item_path(t.commentable.commentable)
           end
         end
         column "Time" do |t|
@@ -100,8 +99,38 @@ ActiveAdmin::Dashboards.build do
         column "IP" do |t|
           link_to(t.user_ip, "http://www.geoiptool.com/en/?IP=#{t.user_ip}", target: "_blank")
         end
-        column "Spam" do |t|
-          t.marked_spam
+        column "Action" do |t|
+          link_to "Delete", admin_comment_path(t), method: 'delete', confirm: "Are you sure you want to delete this comment?", remote: true
+        end
+      end
+    end
+  end
+
+  section "Spam Comments", priority: 2 do
+    if controller.current_ability.can?(:read, Comment)
+      table_for Comment.as_spam(12) do
+        column "User" do |t|
+          if t.owner
+            link_to t.owner.title, admin_user_path(t.owner), class: "suspicious_#{t.suspicious?} spam_#{t.marked_spam?}"
+          else
+            t.owner_id
+          end
+        end
+        column "Message" do |t|
+          link_to t.body.truncate(80), admin_comment_path(t), title: t.body, class: "suspicious_#{t.suspicious?} spam_#{t.marked_spam?}"
+        end
+        column "Item" do |t|
+          if t.commentable && t.commentable_type == "Item"
+            link_to t.commentable.title.truncate(44), admin_item_path(t.commentable)
+          elsif t.commentable && t.commentable_type == "Comment"
+            link_to t.commentable.commentable.title.truncate(44), admin_item_path(t.commentable.commentable)
+          end
+        end
+        column "Time" do |t|
+          "#{time_ago_in_words(t.created_at)} ago"
+        end
+        column "IP" do |t|
+          link_to(t.user_ip, "http://www.geoiptool.com/en/?IP=#{t.user_ip}", target: "_blank")
         end
         column "Action" do |t|
           link_to "Delete", admin_comment_path(t), method: 'delete', confirm: "Are you sure you want to delete this comment?", remote: true
